+++ src/main/java/com/skytechautosieve/sieves/TileEntityAutoSieve.java
@@ -46,7 +46,7 @@ public class TileEntityAutoSieve extends TileEntity implements ITickable, IInven
 	static {
 		Object processTime = InternalTools.readConfig().get(InternalTools.CONFIG_PROCESS_TIME);
 		if (processTime != null) {
-			PROCESS_TIME_REQUIRED = Integer.parseInt(processTime.toString()) * 20;
+			PROCESS_TIME_REQUIRED = (int) Double.parseDouble(processTime.toString()) * 20;
 		}
 	}
